Incoporate iP code quality feedback from automated anaylsis
There were several inconsistencies with code quality which were fixed
to follow the module's coding standard.

Following the coding standard will lead to better code readability,
and possible higher marks for my iP in this module.
