+++ src/main/java/duke/tasks/TaskList.java
@@ -1,4 +1,4 @@
-package duke.Tasks;
+package duke.tasks;
 
 import java.util.ArrayList;
